server/requirements.txt
# sha256: vVfZUHeNuB9V-J78vLkF7oOad4unkK5DCLijFoNet84
# sha256: O7YFNrL7IIRhL8lIZjQpXnIIeQApCBhCUkkWtaZtIG8
Django==1.8.2
# sha256: 8uJz7TSsu1YJYtXPEpF5NtjfAil98JvTCJuFRtRYQTg
# sha256: ygF2j97N4TQwHzFwdDIm9g7f9cOTXxJDc3jr2RFQY1M
dj-database-url==0.3.0
# sha256: ufv3R_qOcR0zDy9Da5suz17tnrZ6Y3yoGrzRHCq-3sg
python-decouple==2.3
# sha256: as-au751fvddwuzZ2RunSVR5Qauv--af8ghqnjfUkEw
psycopg2==2.6.1
# sha256: fn9zpnXFGHErrdeDJ54m0WQUDz_C7XoyECw9CKaipKc
jsonfield==1.0.3
# sha256: J_-iz0sFAeyfxBDyUmKmh3dtO2Qy0lEX8vt69Bia1Kw
# sha256: YPJ9jdB0c4sT7ZJqy7C9UbTnFn05Q6BFZkVxGuIrGrQ
djangorestframework==3.1.3
# sha256: IPl2zc4CpCtpzoDp4DiXpRgUs21EizcohUYIbrxHMUY
# sha256: OYo9ttYYmdJf1KBsbKEgUbDOFx1wXezX7VURUXtLuT0
requests==2.7.0
# sha256: 3yHRQxQkxRKQHsw80kQDmDOuvEvH00WvQD1KvaziAIE
# sha256: Z-HBi2uZykwsjkoC4KlhdL9waJGaWv9Sg2YAl0we9Ng
python-memcached==1.54
# sha256: QYqTw5en7asj5ViNvAZ6x0pyPts9VBvUk295R252Rdo
# sha256: 4kBSQR_E-9H2cmNVN8P8IzDZSBsYwDF2lbRiWVEskdU
six==1.9.0
# sha256: FbWCF9HFWn2BlKhrkDbOHUuItyx0CdlXWjxRXcCVBJA
arrow==0.5.4
# sha256: PpVEXB21AKNEB5pHsXHEXvGPV9GI3_2w5BZccb6o6z0
# sha256: KuY89HXwvQSbci-sIIE9Yq7cFJV91aO_ANEg0rVARGA
python-dateutil==2.4.2
